+++ apps/web/src/App.tsx
@@ -1,4 +1,5 @@
 import { Routes, Route } from 'react-router-dom';
+import WorkspaceLayout from './components/WorkspaceLayout';
 import Home from './pages/Home';
 import Space from './pages/Space';
 import Logto from './pages/Logto';
@@ -8,8 +9,8 @@ export default function App() {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
-      <Route path="/space" element={<Space />} />
-      <Route path="/space/:id?" element={<Space />} />
+      <Route path="/space" element={<WorkspaceLayout><Space /></WorkspaceLayout>} />
+      <Route path="/space/:id?" element={<WorkspaceLayout><Space /></WorkspaceLayout>} />
       <Route path="/logto" element={<Logto />} />
       <Route path="/logto-callback" element={<LogtoCallback />} />
     </Routes>