+++ apps/web/webpack.config.js
@@ -14,7 +14,7 @@ module.exports = (_, { mode }) => ({
   module: {
     rules: [
       { test: /\.tsx?$/, use: { loader: 'ts-loader', options: { transpileOnly: true } }, exclude: /node_modules/ },
-      { test: /\.css$/, use: ['style-loader', 'css-loader'] },
+      { test: /\.css$/, use: ['style-loader', 'css-loader', 'postcss-loader'] },
     ],
   },
   plugins: [